unsigned is more saner than double for binheap items' key

+++ include/binary_heap.h
@@ -93,13 +93,13 @@ struct binheap2 *
 binheap2_new(void);
 
 struct binheap2_item *
-binheap2_insert(struct binheap2 *bh2, void *p, double key);
+binheap2_insert(struct binheap2 *bh2, void *p, unsigned key);
 
 void
 binheap2_delete(struct binheap2 *bh2, struct binheap2_item *bi);
 
 void
-binheap2_reorder(struct binheap2 *bh2, struct binheap2_item *bi, double key);
+binheap2_reorder(struct binheap2 *bh2, struct binheap2_item *bi, unsigned key);
 
 void *
 binheap2_root(struct binheap2 *bh2);
